[kernel] rename to kernel_common.hpp
