+++ hammerblade/torch/kernel/kernel.cpp
@@ -3,7 +3,7 @@
 //====================================================================
 // This file also holds the tile group barrier definition
 
-#include <common.h>
+#include <kernel_common.hpp>
 
 // INIT_TILE_GROUP_BARRIER(r_barrier, c_barrier, 0, bsg_tiles_X-1,
 //     0, bsg_tiles_Y-1);
