+++ hammerblade/torch/kernel/kernel_common.hpp
